Изменен тип поля price на decimal

+++ OLT/materials/models.py
@@ -8,7 +8,7 @@ class Course(models.Model):
     pict = models.ImageField(upload_to='course/', verbose_name='превью', null=True, blank=True)
     description = models.CharField(max_length=250, verbose_name='описание', null=True, blank=True)
     user = models.ForeignKey(User, on_delete=models.DO_NOTHING, default=None, verbose_name='user_id')
-    price = models.PositiveIntegerField(default=0, verbose_name='Цена курса')
+    price = models.DecimalField(decimal_places=2, max_digits=10, default=0, verbose_name='Цена курса')
 
     def __str__(self):
         return f'{self.title}'
@@ -25,7 +25,7 @@ class Lesson(models.Model):
     url_video = models.URLField(verbose_name='видео ссылка', blank=True, null=True)
     course = models.ForeignKey(Course, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.DO_NOTHING, default=None, verbose_name='user_id')
-    price = models.PositiveIntegerField(default=0, verbose_name='Цена урока')
+    price = models.DecimalField(decimal_places=2, max_digits=10, default=0, verbose_name='Цена урока')
 
     def __str__(self):
         return f'{self.title}'
